Update pack and unpack to be able to detect runtime errors
This commit makes a breaking API CHANGE! ! ! ! ! !
dbus message no longer supports streaming operators for unpacking and
packing.
This behavior seemed odd in the first place, and lead to some poor API
issues.  It was also redundant to the much more used messsage.pack
which had far more features, and was much more useable.  message.pack
now returns a bool, indicating whether or not the unpack/pack succeeded.
properties API is also updated to use this new interface to its
advantage, allowing badly formatted pack statements (which should be
expected at runtime) to not require throwing an exception to handle.

Change-Id: I8da22e19106f3c764acfea3db13adc974aa2cb9f